python_app/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
Pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
stripe==9.9.0
ics==0.7.2
gunicorn==22.0.0
requests
parsel
extruct
w3lib
python-dateutil
PyYAML